Created dropdown component, corrected click-event for closing dropdown on click outside of container
